+++ Sourcery/Models/TemplateContext.swift
@@ -1,0 +1,163 @@
+//
+// Created by Krzysztof Zablocki on 31/12/2016.
+// Copyright (c) 2016 Pixle. All rights reserved.
+//
+
+import Foundation
+
+final class TemplateContext: NSObject, SourceryModel {
+    let types: Types
+    let arguments: [String: NSObject]
+
+    init(types: Types, arguments: [String: NSObject]) {
+        self.types = types
+        self.arguments = arguments
+    }
+
+    // sourcery:inline:TemplateContext.AutoCoding
+        /// :nodoc:
+        required internal init?(coder aDecoder: NSCoder) {
+            guard let types: Types = aDecoder.decode(forKey: "types") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["types"])); fatalError() }; self.types = types
+            guard let arguments: [String: NSObject] = aDecoder.decode(forKey: "arguments") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["arguments"])); fatalError() }; self.arguments = arguments
+        }
+
+        /// :nodoc:
+        internal func encode(with aCoder: NSCoder) {
+            aCoder.encode(self.types, forKey: "types")
+            aCoder.encode(self.arguments, forKey: "arguments")
+        }
+    // sourcery:end
+
+    var stencilContext: [String: Any] {
+        return [
+            "types": types,
+            "type": types.typesByName,
+            "argument": arguments
+        ]
+    }
+
+    // sourcery: skipDescription, skipEquality
+    var jsContext: [String: Any] {
+        return [
+            "types": [
+                "all": types.all,
+                "protocols": types.protocols,
+                "classes": types.classes,
+                "structs": types.structs,
+                "enums": types.enums,
+                "based": types.based,
+                "inheriting": types.inheriting,
+                "implementing": types.implementing
+            ],
+            "type": types.typesByName,
+            "argument": arguments
+        ]
+    }
+
+}
+
+// sourcery: skipJSExport
+/// Collection of scanned types for accessing in templates
+public final class Types: NSObject, SourceryModel {
+    let types: [Type]
+
+    init(types: [Type]) {
+        self.types = types
+    }
+
+    // sourcery:inline:Types.AutoCoding
+        /// :nodoc:
+        required public init?(coder aDecoder: NSCoder) {
+            guard let types: [Type] = aDecoder.decode(forKey: "types") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["types"])); fatalError() }; self.types = types
+        }
+
+        /// :nodoc:
+        public func encode(with aCoder: NSCoder) {
+            aCoder.encode(self.types, forKey: "types")
+        }
+    // sourcery:end
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// :nodoc:
+    public lazy internal(set) var typesByName: [String: Type] = {
+        var typesByName = [String: Type]()
+        self.types.forEach { typesByName[$0.name] = $0 }
+        return typesByName
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known types, excluding protocols
+    public lazy internal(set) var all: [Type] = {
+        return self.types.filter { !($0 is Protocol) }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known protocols
+    public lazy internal(set) var protocols: [Protocol] = {
+        return self.types.flatMap { $0 as? Protocol }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known classes
+    public lazy internal(set) var classes: [Class] = {
+        return self.all.flatMap { $0 as? Class }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known structs
+    public lazy internal(set) var structs: [Struct] = {
+        return self.all.flatMap { $0 as? Struct }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known enums
+    public lazy internal(set) var enums: [Enum] = {
+        return self.all.flatMap { $0 as? Enum }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// Types based on any other type, grouped by its name, even if they are not known.
+    /// `types.based.MyType` returns list of types based on `MyType`
+    public lazy internal(set) var based: [String: [Type]] = {
+        var content = [String: [Type]]()
+        self.types.forEach { type in
+            type.based.keys.forEach { name in
+                var list = content[name] ?? [Type]()
+                list.append(type)
+                content[name] = list
+            }
+        }
+        return content
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// Classes inheriting from any known class, grouped by its name.
+    /// `types.inheriting.MyClass` returns list of types inheriting from `MyClass`
+    public lazy internal(set) var inheriting: [String: [Type]] = {
+        var content = [String: [Type]]()
+        self.classes.forEach { type in
+            type.inherits.keys.forEach { name in
+                var list = content[name] ?? [Type]()
+                list.append(type)
+                content[name] = list
+            }
+        }
+        return content
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// Types implementing known protocol, grouped by its name.
+    /// `types.implementing.MyProtocol` returns list of types implementing `MyProtocol`
+    public lazy internal(set) var implementing: [String: [Type]] = {
+        var content = [String: [Type]]()
+        self.types.forEach { type in
+            type.implements.keys.forEach { name in
+                var list = content[name] ?? [Type]()
+                list.append(type)
+                content[name] = list
+            }
+        }
+        return content
+    }()
+
+}